supported_models.py
import google.generativeai as genai
import os

# PASTE YOUR KEY HERE
api_key = "YOUR_GEMINI_API_KEY_HERE"
genai.configure(api_key=api_key)

print("Checking available models...")
try:
    for m in genai.list_models():
        if 'generateContent' in m.supported_generation_methods:
            print(f"- {m.name}")
except Exception as e:
    print(f"Error listing models: {e}")