requirements.txt
pandas
openpyxl
google-generative-ai
google-genai